--- src/main/java/Milo/Tasks/TaskType.java
@@ -1,15 +1,0 @@
-package Milo.Tasks;
-
-/*
- * Represents the different possible task type of the task object
- */
-public class TaskType {
-    public static enum taskType {
-        TODO,
-        EVENT,
-        DEADLINE,
-        INVALID,
-
-        DATE
-    }
-}
