+++ src/main/java/milo/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Milo.Tasks;
+package milo.tasks;
 import java.time.LocalDate;
 
 /**
@@ -9,11 +9,11 @@ import java.time.LocalDate;
 public class Deadline extends Task {
     private final LocalDate deadline;
 
-    /*
+    /**
     * initialise a deadline task object
     * task status set as not completed
     *
-    * @param description of the deadline task
+    * @param desc of the deadline task
     * @param deadline date
      */
     public Deadline(String desc, LocalDate deadline) {
@@ -21,20 +21,20 @@ public class Deadline extends Task {
         this.deadline = deadline;
     }
 
-    /*
+    /**
      * initialise a deadline task object
      * task status may be set as completed
      *
-     * @param description of the deadline task
+     * @param desc of the deadline task
      * @param deadline date
-     * @param task completion status
+     * @param isCompleted completion status
      */
     public Deadline(String desc, LocalDate deadline, Boolean isCompleted) {
         super(desc, isCompleted);
         this.deadline = deadline;
     }
 
-    /*
+    /**
      * overrides the task toString() method
      * provides additional task type and deadline information
      * format string for Ui
@@ -44,7 +44,7 @@ public class Deadline extends Task {
         return "[D]" + super.toString() + " (by: " + deadline + ")";
     }
 
-    /*
+    /**
      * overrides the task toTextString() method
      * provides additional task type and deadline information
      * format string for Storage
