+++ src/main/java/milo/tasks/Task.java
@@ -1,5 +1,5 @@
-package Milo.Tasks;
-/*
+package milo.tasks;
+/**
  * Represents a task event containing
  * completion status, deletion status and description
  * also contains a static count of the task number
@@ -10,47 +10,47 @@ public class Task {
 
     private Boolean isDeleted = false;
 
-    /*
+    /**
      * initialise a task object
      * task status set as not completed
      *
-     * @param description of the task
+     * @param desc of the task
      */
     public Task(String desc) {
         this.description = desc;
     }
 
-    /*
+    /**
      * initialise a task object
      * task status may be set as completed
      *
-     * @param description of the task
-     * @param task completion status
+     * @param desc of the task
+     * @param isCompleted completion status
      */
     public Task(String desc, Boolean isCompleted) {
         this.description = desc;
         this.isCompleted = isCompleted;
     }
 
-    /*
+    /**
     * mark task completion status as completed
-     */
+    */
     public void mark() {
         if (!this.isCompleted) {
             this.isCompleted = true;
         }
     }
 
-    /*
-     * mark task completion status as incomplete
-     */
+    /**
+    * mark task completion status as incomplete
+    */
     public void unmark() {
         if (this.isCompleted) {
             this.isCompleted = false;
         }
     }
 
-    /*
+    /**
      * deletes current task
      * decrementing taskNumber by one
      */
@@ -62,7 +62,7 @@ public class Task {
         return this.description != null && this.description.contains(taskDesc);
     }
 
-    /*
+    /**
      * overrides the Object toString() method
      * provides completion and description information
      * format string for Ui
@@ -72,7 +72,7 @@ public class Task {
         return isCompleted ? "[X] " + description : "[ ] " + description;
     }
 
-    /*
+    /**
      * return different formatted string information of the task
      * provides completion and description information
      * format string for Storage
