+++ src/test/java/milo/UiTest.java
@@ -1,9 +1,14 @@
-package Milo;
-import Milo.Ui.Ui;
-import org.junit.jupiter.api.Test;
+package milo;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+import milo.ui.Ui;
+
+
 
 class UiTest {
     @Test
