+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Milo
+Hello! I'm milo
 What can I do for you?
   ╱|、
 (˚ˎ 。7
